refactor move ModActionMenu to modListingPanel file
